+++ packages/jxl/codec/pre.js
@@ -1,4 +1,4 @@
-const isRunningInCloudflareWorker = caches.default !== undefined;
+const isRunningInCloudflareWorker = globalThis.caches && globalThis.caches.default !== undefined;
 
 if (isRunningInCloudflareWorker) {
   if (!globalThis.ImageData) {